addTooltip refractor to ToolBarItem
