+++ lib/items/tool-bar-button-view.js
@@ -1,6 +1,5 @@
 const {CompositeDisposable} = require('atom');
-const { ToolBarItem } = require('./tool-bar-item');
-
+const { ToolBarItem, addTooltip } = require('./tool-bar-item');
 class ToolBarButtonView extends ToolBarItem {
   constructor (options, group) {
     // first calling the super (ToolBarItem) constructor
@@ -9,34 +8,9 @@ class ToolBarButtonView extends ToolBarItem {
       priority: options.priority
     }, group);
 
-    this.subscriptions = new CompositeDisposable();
     this.options = options;
     this.enabled = true;
 
-    if (options.tooltip) {
-      const callback = this.options.callback;
-
-      let tooltip = {};
-      if (typeof options.tooltip === 'string') {
-        tooltip = {
-          title: options.tooltip
-        };
-      } else {
-        tooltip = options.tooltip;
-      }
-
-      if (!tooltip.hasOwnProperty('placement')) {
-        tooltip.placement = getTooltipPlacement;
-      }
-
-      if (!tooltip.hasOwnProperty('keyBindingCommand')) {
-        tooltip.keyBindingCommand = typeof callback === 'string' ? callback : null;
-      }
-
-      this.subscriptions.add(
-        atom.tooltips.add(this.element, tooltip)
-      );
-    }
 
     const classNames = ['btn', 'btn-default', 'tool-bar-btn'];
     if (this.priority < 0) {
@@ -60,6 +34,11 @@ class ToolBarButtonView extends ToolBarItem {
       } else {
         this.element.textContent = options.text;
       }
+    if (options.tooltip) {
+      this.subscriptions = new CompositeDisposable();
+      this.subscriptions.add(
+        addTooltip(this.options.tooltip, this.options.callback)
+      );
     }
 
     if (options.color) {
@@ -156,15 +135,6 @@ class ToolBarButtonView extends ToolBarItem {
       }
     }
   }
-};
-
-function getTooltipPlacement () {
-  const toolbarPosition = atom.config.get('tool-bar.position');
-  return toolbarPosition === 'Top' ? 'bottom'
-    : toolbarPosition === 'Right' ? 'left'
-      : toolbarPosition === 'Bottom' ? 'top'
-        : toolbarPosition === 'Left' ? 'right'
-          : null;
 }
 
 function getCallbackModifier (callback, {altKey, ctrlKey, shiftKey}) {
